Added a database major changes
For now, only the 'Talking' mechanic works with this .db file. The
database consists of three columns, 'Dialogos' (who stores dialogues),
'Relaciones' (who stores relationships with other characters), and
'DialogosRelaciones' (an intermediate table for changing the Dialogues
depending on the relationship with the other character).
Added three classes (.cs) for now, ClsConexion (who allows you to
connect with the .db file in any class), ClsDialogo (who stores the Dialogos
table for controlling it), and ClsDialogos (who manages the connection with
the table).
Talking with the girl now shows you the dialogues from the .db.